==> scripts/05_previous_work.Rmd <==
---
title: "Potential workstationns"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r}
library(raster)
library(sf)
library(tidyverse)
library(ggspatial)

ps_data_path <- "/Volumes/ps-science/datasets/"

ps_exp_path <- "/Volumes/Colombia - 2022/"

default_font_family <- "Arial"

default_font_color <- '#22211d'

default_background_color <- "#f5f5f2"

load(file = file.path(ps_exp_path, "data", "secondary", "processed", "admin_spatial_layers.RData"))

theme_map <- function(...) {
  theme_minimal() +
    theme(
      text = element_text(family = default_font_family,
                          color = default_font_color),
      # remove all axes
      # add a subtle grid
      panel.grid.major = element_line(color = "#dbdbd9", size = 0.2),
      panel.grid.minor = element_blank(),
      # background colors
      plot.background = element_rect(fill = default_background_color,
                                     color = NA),
      panel.background = element_rect(fill = default_background_color,
                                      color = NA),
      legend.background = element_rect(fill = default_background_color,
                                       color = NA),
      # borders and margins
      plot.margin = unit(c(.5, .5, .2, .5), "cm"),
      panel.border = element_blank(),
      panel.spacing = unit(c(-.1, 0.2, .2, 0.2), "cm"),
      # titles
      legend.title = element_text(size = 11),
      legend.text = element_text(size = 9, hjust = 0,
                                 color = default_font_color),
      plot.title = element_text(size = 15, hjust = 0.5,
                                color = default_font_color),
      plot.subtitle = element_text(size = 10, hjust = 0.5,
                                   color = default_font_color,
                                   margin = margin(b = -0.1,
                                                   t = -0.1,
                                                   l = 2,
                                                   unit = "cm"),
                                   debug = F),
      # captions
      plot.caption = element_text(size = 7,
                                  hjust = .5,
                                  margin = margin(t = 0.2,
                                                  b = 0,
                                                  unit = "cm"),
                                  color = "#939184"),
      ...
    )
}
```

# Offshore Pacific

## Argo Navigator 

```{r}
navigator_depths <- readxl::read_xlsx(file.path(ps_exp_path, 
                                                "data", "secondary", "raw",  "Argo_navigator_readings_2021.xlsx")) %>% 
  janitor::clean_names() %>% 
  rename(c(lon = "w", lat = "n")) 

navigator_depths$lat_d = as.numeric(measurements::conv_unit(navigator_depths$lat, from = 'deg_dec_min', to = 'dec_deg'))

navigator_depths$lon_d = as.numeric(measurements::conv_unit(navigator_depths$lon, from = 'deg_dec_min', to = 'dec_deg'))

navigator_depths <- navigator_depths %>% 
  mutate(depth_m = as.numeric(depth_m)) %>% 
  dplyr::select(point_number, lon_d, lat_d, depth_m) %>% 
  mutate(lon_d = -1*lon_d)

navigator_depths <- navigator_depths %>% 
  na.omit() %>% 
  arrange(desc(lon_d)) %>% 
  head(-1) %>%  
  mutate(first_site = if_else(point_number == "089", T, F))

navigator_depths_sf <- navigator_depths %>% 
  sf::st_as_sf(coords = c("lon_d", "lat_d"), 
               crs = 4326) %>% 
  sf::st_transform(crs = col_crs) 

navigator_depths_sf %>% 
  sf::st_write(file.path(ps_exp_path, "data", "secondary", "processed",  "argo_navigator_depths.gpkg"))
```

## INVEMAR drop stations in Colinas y Lomas

```{r}
invemar_drops <- read_csv(file.path(ps_exp_path, "data", "secondary", "raw", "invemar_drop_coords.csv"))

invemar_drops_sf <- invemar_drops %>% 
    st_as_sf(coords = c("longitud", "latitud"),
           crs = 4326) %>% 
  st_transform(crs = col_crs) 

invemar_drops_sf %>% 
  sf::st_write(file.path(ps_exp_path, "data", "secondary", "processed", "invemar_drops.gpkg"))

# invemar_drops %>% 
#   mutate(lon = measurements::conv_unit(longitud,
#                                        to = 'deg_dec_min', 
#                                        from = 'dec_deg'),
#          lat = measurements::conv_unit(latitud,
#                                        to = 'deg_dec_min', 
#                                        from = 'dec_deg'))
```

# Tribuga

## INVEMAR/ANH stations

```{r}
invemar_tribuga_stations <- read_csv(file.path(ps_exp_path, 
                                       "data", "secondary", "raw", "estaciones_AUNAP_INVEMAR.csv"))

# change the degree symbol to a space
invemar_tribuga_stations$lat = gsub('°', ' ', invemar_tribuga_stations$lat)
invemar_tribuga_stations$lat = gsub(',', '.', invemar_tribuga_stations$lat)

invemar_tribuga_stations$lon = gsub('°', ' ', invemar_tribuga_stations$lon)
invemar_tribuga_stations$lon = gsub(',', '.', invemar_tribuga_stations$lon)

# convert from decimal minutes to decimal degrees
invemar_tribuga_stations$lat_d = as.numeric(measurements::conv_unit(invemar_tribuga_stations$lat, 
                                                                    from = 'deg_dec_min', 
                                                                    to = 'dec_deg'))

invemar_tribuga_stations$lon_d = as.numeric(measurements::conv_unit(invemar_tribuga_stations$lon, 
                                                                    from = 'deg_dec_min', 
                                                                    to = 'dec_deg'))

invemar_tribuga_stations %>% 
  select(aunap_station_id, station_name, lat_d, lon_d) %>% 
  mutate(lon_d = -1*lon_d) %>% 
  sf::st_as_sf(coords = c("lon_d", "lat_d"), crs = 4326) %>% 
  st_transform(crs = st_crs(bajos_eez)) %>% 
  st_write(file.path(project_path, "data", "02_processed_data", "invemar_tribuga_stations.gpkg"), append = F)
```

## Puntos Sandra

```{r}
pts_sandra <- read_csv(file = file.path(ps_exp_path, 
                                       "data", "secondary", "raw", "puntos_golfo_sandra.csv")) %>% 
  arrange(desc(lat))

pts_sandra <- pts_sandra %>% 
  mutate(lon = paste0("-", lon),
         lon_dec = as.numeric(measurements::conv_unit(lon,
                                           from = "deg_min_sec",
                                           to = "dec_deg")),
         lat_dec = as.numeric(measurements::conv_unit(lat,
                                           from = "deg_min_sec",
                                           to = "dec_deg"))) %>% 
  mutate_if(is.numeric,round, 4) %>% 
  select(source, name = nombre, lon, lat, lon_dec, lat_dec)

pts_sandra <- pts_sandra %>% 
  mutate(lon = measurements::conv_unit(lon,
                                       from = "deg_min_sec",
                                       to = "deg_dec_min"),
         lat = measurements::conv_unit(lat,
                                       from = "deg_min_sec",
                                       to = "deg_dec_min")) %>% 
  mutate(lon = stringr::str_trunc(lon, 8, "right", ellipsis = ""),
         lat = stringr::str_trunc(lat, 8, "right", ellipsis = ""))

pts_sandra %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(ps_exp_path, 
                     "data", "secondary", "processed",  "pts_sandra.gpkg"), append = F)
```

## Turqui

```{r}
pts_turqui <- read_csv(file.path(ps_exp_path, "data", "secondary", "raw","turqui_pts.csv")) 

pts_turqui <- pts_turqui %>% 
  rename(depth_m = depth) %>% 
  mutate(source = "turqui",
         lon_dec = as.numeric(measurements::conv_unit(lon,
                                                      from = "deg_dec_min",
                                                      to = "dec_deg")),
         lat_dec = as.numeric(measurements::conv_unit(lat,
                                                      from = "deg_dec_min",
                                                      to = "dec_deg")))
pts_turqui %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(ps_exp_path, "data", "secondary", "processed", "pts_turqui.gpkg"), append = F)
```

# Bajo Nuevo & Serrranilla

```{r}
estaciones_vega <- read_csv(file.path(ps_exp_path, "data", "secondary", "raw", "estaciones_seaflower_Vega_2016.csv"))

estaciones_vega <- estaciones_vega %>% 
  sf::st_as_sf(coords = c("Long", "Lat"), crs = 4326) %>% 
  st_transform(crs = st_crs(bajos_eez))

estaciones_vega <- estaciones_vega %>% 
  st_join(bajos_eez %>% select(territory1), left = T) %>% 
  rename(roi = territory1) %>% 
  replace_na(list(roi = "Bajo Alicia")) 

estaciones_vega %>% 
  st_write(file.path(ps_exp_path, "data", "secondary", "processed",  "vega_stations.gpkg"))

estaciones_vega %>% 
  st_transform(crs = 4326) %>% 
  dplyr::mutate(lon = sf::st_coordinates(.)[,1],
                lat = sf::st_coordinates(.)[,2]) %>% 
  select(lon, lat, roi ) %>% 
  st_drop_geometry() %>% 
  mutate_at(vars(lon, lat), round, 5) %>% 
  write_csv(file.path(ps_exp_path, "data", "secondary", "processed", "vega_stations.csv"))
```

## AGGRA stations

```{r}
aggra_stations <- read_csv(file.path(ps_exp_path, "data", "secondary", "raw", "AGRRA_stations_seaflower.csv"))

aggra_stations <- aggra_stations %>% 
  sf::st_as_sf(coords = c("Lon", "Lat N"), crs = 4326) %>% 
  select(-"Long W") %>% 
  st_transform(crs = st_crs(bajos_eez))

aggra_stations %>% 
  st_write(file.path(ps_exp_path, "data", "secondary", "processed",  "aggra_stations.gpkg"), append = F)

aggra_stations %>% 
  st_transform(crs = 4326) %>% 
  dplyr::mutate(lon = sf::st_coordinates(.)[,1],
                lat = sf::st_coordinates(.)[,2]) %>% 
  janitor::clean_names() %>% 
  st_drop_geometry() %>% 
  mutate_at(vars(lon, lat), round, 5) %>% 
  write_csv(file.path(ps_exp_path, "data", "secondary", "processed", "aggra_stations.csv"))
```
