==> scripts/08_analyze_uvc_fish.Rmd <==
---
title: "Reef fish communities"
output:
  word_document:
    toc: yes
    toc_depth: '1'
  html_document:
    toc: yes
    toc_depth: '1'
    df_print: paged
---

```{r echo = FALSE, include = FALSE, warning = FALSE,prompt = FALSE, error = FALSE, strip.white = TRUE}
knitr::opts_chunk$set(echo = FALSE, include = FALSE, warning = FALSE,
                      prompt = FALSE, error = FALSE, strip.white = TRUE)

knitr::opts_knit$set(progress = TRUE, verbose = TRUE)

options(dplyr.summarise.inform = FALSE)

library(sf)
library(tidyverse)
library(vegan)

source("_prj_vars.R")

load(file = file.path(ps_exp_path, 
                      "data", "secondary", "processed", "admin_spatial_layers.RData"))

uvc_meta <- read_csv(file.path(ps_exp_path, 
                               "data", "primary", "processed", "metadata", "uvs_meta.csv")) %>% 
  filter(method == "Fish survey", location == "Caribbean") %>% 
  mutate(habitat = str_extract(habitat, "[^,]+"))

uvc_taxa <- read_csv(file.path(ps_exp_path, 
                               "data", "primary", "processed", "taxa", "clean_master_taxa.csv"))%>% 
  filter(method == "uvs")

caribe_fish_data <- readxl::read_xlsx(file.path(ps_exp_path, 
                               "data", "primary","raw", "fish", "SEA_fish_final.xlsx"))
```

```{r}
uvc_meta %>% 
  group_by(sublocation, habitat) %>% 
  summarize(n_distinct(ps_station_id))
```

```{r}
caribe_sat_img <- terra::rast(file.path(ps_exp_path, 
                               "data", "secondary", "raw", 
                               "Allen atlas", "Coral-Reefs-2020-Visual-V1-Mosaic", "satellite_imagery_0.tif")) 

bajos_eez <- bajos_eez %>% 
  st_transform(terra::crs(caribe_sat_img)) 
  
bn_sat_img <- caribe_sat_img %>% 
  terra::crop(bajos_eez %>% 
                filter(territory1 == "Bajo Nuevo Bank") %>% 
                terra::ext(), mask = T) 

serr_sat_img <- caribe_sat_img %>% 
  terra::crop(bajos_eez %>% 
                filter(territory1 == "Serranilla Bank") %>% 
                terra::ext(), mask = T) 
```

# Clean data 

```{r}
caribe_fish_data <- caribe_fish_data %>%
  rename(min_length = min, max_length = max, n_ind = number, ps_taxon_code = species) %>%
  rowwise() %>% 
  mutate(avg_length = (min_length + min_length)/2,
         depth_strata = cut(depth_m, breaks = c(0, 15, Inf), labels = c("S", "D"))) %>% 
  dplyr::select(SEA_station_id = ps_station_id, depth_strata, depth_m, diver, transect, 
         ps_taxon_code, n_ind,avg_length) %>% 
  left_join(uvc_meta %>% 
              filter(location == "Caribbean") %>% 
              distinct(ps_station_id, sublocation, habitat) %>% 
              mutate(SEA_station_id = paste0("SEA_fish_", 
                                             formatC(row_number(), width = 2, flag = 0)))) %>% 
  dplyr::select(sublocation, ps_station_id, everything(), -SEA_station_id) %>% 
  left_join(uvc_taxa %>% 
              filter(location == "Caribbean") %>% 
              distinct(ps_taxon_code, taxon_sci_name, family, order, class,
                       is_vme, iucn_redlist_cat, max_length_tl, troph, a, b)) 
## QA/QC

caribe_fish_data %>%
  filter(avg_length > max_length_tl) %>%
  transmute(abs(avg_length - max_length_tl))

caribe_fish_data <- caribe_fish_data %>% 
  mutate(avg_length = if_else(avg_length > max_length_tl & !is.na(max_length_tl),
                              max_length_tl, 
                              avg_length))
```

## By transect 

```{r}
fish_data_by_transect <- caribe_fish_data %>% 
  mutate(transect_id = paste(ps_station_id, transect, diver, depth_strata, sep = "_"),
         biomass_gr = n_ind*a*avg_length^b,
         ind_m2 = if_else(avg_length <= 20, n_ind/50, n_ind/100),
         gr_m2 = if_else(avg_length <= 20, biomass_gr/50, biomass_gr/100)) %>% 
  group_by(sublocation, ps_station_id, transect_id, transect, diver, depth_strata, depth_m, habitat,
           ps_taxon_code, taxon_sci_name, order, family, iucn_redlist_cat, troph) %>% 
  summarise(across(c(n_ind, ind_m2, biomass_gr, gr_m2), sum)) %>% 
  ungroup() 

# Need to add zeros

fish_data_by_transect <- fish_data_by_transect %>% 
  distinct(sublocation, ps_station_id, transect_id, diver, depth_strata, depth_m, habitat) %>% 
  full_join(expand.grid(transect_id = unique(fish_data_by_transect$transect_id),
                        ps_taxon_code = unique(fish_data_by_transect$ps_taxon_code))) %>% 
  left_join(fish_data_by_transect %>% 
              distinct(transect_id, ps_taxon_code, n_ind, ind_m2, gr_m2)) %>% 
  left_join(fish_data_by_transect %>% 
              distinct(ps_taxon_code, taxon_sci_name, order, family, iucn_redlist_cat)) %>% 
  replace_na(list(n_ind = 0, ind_m2 = 0, gr_m2 = 0))
```


```{r}
transect_summary <- fish_data_by_transect %>% 
  group_by(sublocation, ps_station_id, transect_id, habitat, depth_strata, depth_m, diver) %>% 
  summarize(n_taxa = n_distinct(ps_taxon_code[n_ind > 0]),
            n_ind = sum(n_ind),
            ind_m2 = sum(ind_m2),
            gr_m2 = sum(gr_m2)) %>% 
  ungroup() 

n_ind_data_wide <- fish_data_by_transect %>% 
  dplyr::select(transect_id, ps_taxon_code, n_ind) %>% 
  pivot_wider(names_from = ps_taxon_code, values_from = n_ind, values_fill = 0) %>% 
  column_to_rownames("transect_id") 

diversity_by_transect <- diversity(n_ind_data_wide) %>% 
  enframe() %>% 
  set_names(c("transect_id", "H")) %>% 
  mutate(richness = specnumber(n_ind_data_wide),
         evenness = H/log(richness))

transect_summary <- transect_summary %>% 
  left_join(diversity_by_transect) %>% 
  mutate_if(is.numeric, round, 2) %>% 
  ungroup() 
```

## By station

```{r}
fish_data_by_station <- fish_data_by_transect %>% 
  group_by(sublocation, habitat, ps_station_id, family, taxon_sci_name) %>% 
  summarize(n_ind = mean(n_ind, na.rm = T),
            ind_m2 = mean(ind_m2, na.rm = T),
            gr_m2 = mean(gr_m2, na.rm = T)) %>% 
  ungroup() %>% 
  group_by(ps_station_id) %>% 
  mutate(f_n_ind = n_ind/sum(n_ind),
         f_ind_m2 = ind_m2/sum(ind_m2),
         f_gr_m2 = gr_m2/sum(gr_m2)) %>% 
  ungroup() %>% 
  mutate_if(is.numeric, round, 2) %>% 
  dplyr::select(sublocation, habitat, ps_station_id, family, taxon_sci_name, 
         n_ind, f_n_ind, ind_m2, f_ind_m2, gr_m2, f_gr_m2)

fish_data_by_station %>% 
  write_csv(file.path(ps_exp_path, "reports/Caribe/Anexos", "Tabla_S3_abundancia_y_biomasa_de_peces_por_taxon_por_estacion.csv"))

station_summary <- transect_summary %>% 
  group_by(sublocation, habitat, ps_station_id) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2, H, evenness), mean)) %>% 
  mutate_if(is.numeric, round, 2) %>% 
  ungroup() 

station_summary %>% 
  write_csv(file.path(ps_exp_path, "reports/Caribe/Anexos", "Tabla_S4_metricas_univariadas_de_peces_por_estacion.csv"))

```

```{r}
fish_data_by_station <- read_csv(file.path(ps_exp_path, 
                                           "reports/Caribe/Anexos",                               "Tabla_S3_abundancia_y_biomasa_de_peces_por_taxon_por_estacion.csv"))

fish_data_by_station |> 
  filter(habitat != "seagrass beds") |> 
  filter(family %in% c("Carangidae", "Lutjanidae") |
           str_detect(taxon_sci_name, "Mycteroperca")) |> 
  group_by(sublocation, family, ps_station_id) |> 
  summarize(gr_m2 = sum(gr_m2)) |> 
  ungroup() |> 
  group_by(sublocation, family) |> 
  summarize(gr_m2 = mean(gr_m2))

tmp <- fish_data_by_station |> 
  filter(habitat != "seagrass beds") |> 
  group_by(sublocation, habitat, ps_station_id, family) |> 
  summarize(n_ind = sum(n_ind, na.rm = T),
            ind_m2 = sum(ind_m2, na.rm = T),
            gr_m2 = sum(gr_m2, na.rm = T)) |> 
  ungroup() |> 
  group_by(sublocation, family) |> 
  summarize(n_ind = mean(n_ind, na.rm = T),
            ind_m2 = mean(ind_m2, na.rm = T),
            gr_m2 = mean(gr_m2, na.rm = T)) |> 
  ungroup() |> 
  group_by(sublocation) |> 
  mutate(f_biomass = 100*gr_m2/sum(gr_m2)) 

tmp |> 
  mutate(eat = case_when(family == "Lutjanidae" ~ "Pargos",
                         family == "Serranidae" ~ "Meros",
                         family == "Carangidae" ~ "Jureles",
                         family == "Carcharhinidae" ~ "Tiburones",
                         TRUE ~ NA_character_),
         tmp = case_when(family == "Lutjanidae" ~ T,
                         family == "Serranidae" ~ T,
                         family == "Carangidae" ~ T,
                         family == "Carcharhinidae" ~ T,
                         TRUE ~ F)) |> 
  ggplot(aes(sublocation, gr_m2))+
  geom_col(aes(fill = fct_reorder(family, gr_m2, .na_rm = F),
               alpha = tmp))+
  paletteer::scale_fill_paletteer_d("ggsci::default_igv", direction = -1)+
  scale_alpha_discrete(range = c(0.15, 1))+
  ggthemes::theme_hc()+
  theme(legend.position = "right")+
  labs(fill = "", y = "Biomasa (g/m^2)",
       x = "", 
       title = "Biomasa de peces de arrecife por familia")

ggsave("tst.pdf", height = 7, width = 8 )
```

```{r}
trophic_info <- file.path("~/marine.data.science@ngs.org - Google Drive/My Drive/Pristine Seas/SCIENCE/expeditions/",
          "DMA-2022", "data", "primary", "processed", "uvs_fish_taxa.csv") |> 
  read_csv() |> 
  distinct(taxon_valid_name, trophic_group) |> 
  mutate(trophic_group = case_when(trophic_group %in% c("MI/P", "MI", 
                                                        "SI", "INV") ~ "Invertivores",
                                   trophic_group %in% c("Z") ~ "Planktivores",
                                   trophic_group %in% c("P") ~ "Piscivores",
                                   trophic_group %in% c("H") ~ "Herbivores",
                                   TRUE ~ trophic_group)) 

fish_data_by_station |> 
  left_join(trophic_info, 
            by = c("taxon_sci_name" = "taxon_valid_name")) |> 
  #filter(is.na(trophic_group)) |> 
  distinct(family, taxon_sci_name, trophic_group) |> 
  write_csv("trophic_info.csv")

tmp <- fish_data_by_station |> 
  filter(habitat != "seagrass beds") |> 
  left_join(trophic_info, 
            by = c("taxon_sci_name" = "taxon_valid_name")) |> 
  group_by(sublocation, habitat, ps_station_id, trophic_group) |> 
  summarize(n_ind = sum(n_ind, na.rm = T),
            ind_m2 = sum(ind_m2, na.rm = T),
            gr_m2 = sum(gr_m2, na.rm = T)) |> 
  ungroup() |> 
  group_by(sublocation, trophic_group) |> 
  summarize(n_ind = mean(n_ind, na.rm = T),
            ind_m2 = mean(ind_m2, na.rm = T),
            gr_m2 = mean(gr_m2, na.rm = T)) |> 
  ungroup() |> 
  group_by(sublocation) |> 
  mutate(f_biomass = 100*gr_m2/sum(gr_m2)) 
```

```{r}
parrots_ref <- readxl::read_xlsx("~/marine.data.science@ngs.org - Google Drive/My Drive/Pristine Seas/SCIENCE/datasets/jackson_2014_caribbean_reefs/tbl_2_report.xlsx") |> 
  janitor::clean_names() |> 
  select(territory = countryor_territory, location, parrots_gr_m2 = parrotfishbiomassafter1999_g_m2) |> 
  filter(!is.na(parrots_gr_m2))

parrots_ref |> 
  tibble::add_case(territory = "Bajo Nuevo", 
                   parrots_gr_m2 = 8.7) |> 
  tibble::add_case(territory = "Serranilla", 
                   parrots_gr_m2 = 9.1) |> 
  group_by(territory) |> 
  summarize(parrots_gr_m2 = mean(parrots_gr_m2)) |> 
  ungroup() |> 
  mutate(highlight = if_else(territory %in% c("Bajo Nuevo", "Serranilla"), "yes", "no")) |>
  ggplot(aes(fct_reorder(territory, parrots_gr_m2), 
             parrots_gr_m2, 
             fill = highlight))+
  geom_col( show.legend = F)+
  coord_flip()+
  labs(x = "", y = "Biomasa (g/m2)", title = "Estimativos regionales de biomasa de pez loro")+
  scale_y_continuous(expand = c(0, 0))+
  ggthemes::theme_hc()+
  scale_fill_manual(values = c("yes" = "red", "no" = "grey"))

ggsave("regional_loro.pdf", height = 5, width = 8)
```

```{r}
station_summary <- read_csv(file.path(ps_exp_path, 
                                      "reports/Caribe/Anexos",
                                    "Tabla_S4_metricas_univariadas_de_peces_por_estacion.csv"))
```

```{r, eval = F}
serr_map <- station_summary %>% 
  left_join(uvc_meta %>% 
              distinct(ps_station_id, lat, lon)) %>% 
  filter(sublocation == "Serranilla") %>% 
  sf::st_as_sf(coords = c("lon", "lat"), crs = 4326) %>% 
  ggplot()+
  tidyterra::geom_spatraster_rgb(data = serr_sat_img, maxcell = 10^7)+
  geom_sf(aes(col = habitat, size = gr_m2))+
  coord_sf(xlim = c(-79.86, -79.82), ylim = c(15.785, 15.825))+
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.2, 
                              style = "ticks", 
                              line_col = "white",
                              text_col = "white",
                              pad_y = unit(0.01, units =  "native"), 
                              pad_x = unit(0.01, units =  "native"))+
  ggspatial::annotation_north_arrow(location = "tr", 
                                    which_north = "true", 
                                    height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.01, "native"), pad_y = unit(0.01, "native"),
                                    style = ggspatial::north_arrow_fancy_orienteering)+
  paletteer::scale_color_paletteer_d("RColorBrewer::Set3")+
  scale_shape_manual(values= seq(100:110))+
  theme_light()+
  labs(col = "Habitat", size = bquote('Biomass density '(g/m^2)))+
  guides(size=guide_legend(override.aes=list(colour="grey")))+
  theme(legend.background = element_blank(),
        legend.position = "right",
        #legend.box = 'vertical',
        #legend.justification = 'left',
        #legend.box.just = 'left',
        legend.key = element_blank(),
        legend.text = element_text(colour = "black"), 
        panel.background  = element_rect(fill ="#1b2b49"),
        panel.grid = element_blank())

serr_map_spanish <- serr_map +
    PristineSeasR::scale_color_pristine_seas(palette = "alternative",
                                             labels = c("Terraza prearrecifal", "Arrecife periferico","Cuenca lagunar", "Pastos marinos"))+
    labs(col = "Habitat", size = bquote('Biomasa '(g/m^2)))

ggsave(serr_map_spanish,
       filename = file.path(ps_exp_path, "figures", "serranilla_fish_biomass_map_spanish.png"), 
       width = 8, height = 6,
       dpi = 300)
```

```{r, eval = F}
bn_map <- station_summary %>% 
  left_join(uvc_meta %>% 
              distinct(ps_station_id, lat, lon)) %>% 
  filter(sublocation == "Bajo Nuevo") %>% 
  sf::st_as_sf(coords = c("lon", "lat"), crs = 4326) %>% 
  ggplot()+
  tidyterra::geom_spatraster_rgb(data = bn_sat_img, maxcell = 10^7)+
  geom_sf(aes(col = habitat, size = gr_m2))+
  coord_sf(xlim = c(-78.8, -78.5), ylim = c(15.75, 15.95))+
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.2, 
                              style = "ticks", 
                              line_col = "white",
                              text_col = "white",
                              pad_y = unit(0.01, units =  "native"), 
                              pad_x = unit(0.01, units =  "native"))+
  ggspatial::annotation_north_arrow(location = "tr", 
                                    which_north = "true", 
                                    height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.01, "native"), pad_y = unit(0.01, "native"),
                                    style = ggspatial::north_arrow_fancy_orienteering)+
  #paletteer::scale_color_paletteer_d("RColorBrewer::Set3")+
  scale_shape_manual(values= seq(100:110))+
  theme_light()+
  labs(col = "Habitat", size = bquote('Biomass density'(g/m^2)))+
  guides(size=guide_legend(override.aes=list(colour="grey")))+
  theme(legend.background = element_blank(),
        legend.position = "right",
        # legend.box = 'vertical',
        # legend.justification = 'left',
        # legend.box.just = 'left',
        legend.key = element_blank(),
        legend.text = element_text(colour = "black"), 
        panel.background  = element_rect(fill ="#1b2b49"),
        panel.grid = element_blank())

bn_map_spanish <- bn_map+
  PristineSeasR::scale_color_pristine_seas(palette = "alternative",
                                           labels = c("Terraza prearrecifal profunda", 
                                                "Terraza prearrecifal",
                                                "Cuenca lagunar"))+
  labs(col = "Habitat", size = bquote('Biomasa '(g/m^2)))

ggsave(bn_map_spanish,
       filename = file.path(ps_exp_path, "figures", "bajo_nuevo_fish_biomass_map_spanish.png"), 
       dpi = 300, width = 8, height = 4)
```

```{r}
(location_summary <- station_summary %>% 
   group_by(sublocation) %>% 
   summarize(across(c(richness = n_taxa, n_ind, ind_m2, gr_m2, H), mean)) %>% 
   mutate_if(is.numeric, round, 2) %>% 
   left_join(fish_data_by_transect %>% 
               group_by(sublocation) %>% 
               summarize(n_stations = n_distinct(ps_station_id),
                         n_transects = n_distinct(transect_id),
                         n_families = n_distinct(family[n_ind > 0]),
                         n_taxa = n_distinct(ps_taxon_code[n_ind > 0]))) %>% 
   select(sublocation, n_stations, n_transects, n_taxa, n_families, everything()) %>% 
   ungroup())

(habitat_summary <- station_summary %>% 
    group_by(sublocation, habitat) %>% 
    summarize(across(c(richness = n_taxa, n_ind, ind_m2, gr_m2, H), mean)) %>% 
    mutate_if(is.numeric, round, 2) %>% 
    left_join(fish_data_by_transect %>% 
                group_by(sublocation, habitat) %>% 
                summarize(n_stations = n_distinct(ps_station_id),
                          n_transects = n_distinct(transect_id),
                          n_families = n_distinct(family[n_ind > 0]),
                          n_taxa = n_distinct(ps_taxon_code[n_ind > 0]))) %>% 
    select(sublocation, n_stations, n_transects, n_taxa, n_families, everything()) %>% 
    ungroup() %>% 
    arrange(desc(richness)))
```

```{r}
fish_data_by_transect %>% 
  group_by(order, family) %>% 
  summarise(n_taxa = n_distinct(ps_taxon_code[n_ind > 0])) %>% 
  ungroup() %>% 
  arrange(desc(n_taxa))
```

Overall, we registered `r location_summary$n_taxa[location_summary$sublocation == "Bajo Nuevo"]` and `r location_summary$n_taxa[location_summary$sublocation == "Serranilla"]` distinct taxa from `r location_summary$n_families[location_summary$sublocation == "Bajo Nuevo"]` and `r location_summary$n_families[location_summary$sublocation == "Serranilla"]` families in Bajo Nuevo and Serranilla, respectively.  Families with the most number of taxa were Serranidae (n = 13), Scaridae (n = 11), Pomacentridae (n = 11), and Labridae (n = 10). On average, species richness was slightly higher in Bajo Nuevo (21 taxa per transect) than in Serranilla (18 taxa per transect). This can be partially attributed to the seagrass habitat in Serranilla which had a significantly lower number of species (1.67 species) relative to all other sites (R2 = 0.20; p = 0.001). Differences in species richness between the other habitats were not statistically significant but lagoon patch reefs in Bajo Nuevo had the highest richness (22.80 species), followed by fringing reefs in Serranilla (21.67 species). On average, species richness was higher in shallower transect (20.7 species) than in deeper ones (19.6 species) and even though differences were statistically significant (F = 4.6171; p = 0.001), depth explains little of the variance across transects (R2 = 0.02045). 

```{r, include = T, echo = F}
habitat_summary %>% 
  select(sublocation, habitat, n_stations, n_transects, n_taxa, richness, ind_m2, gr_m2) %>% 
  arrange(sublocation, desc(n_stations)) %>% 
  flextable::flextable() %>% 
  flextable::set_header_labels(values = list(sublocation = "Location", 
                                             habitat = "Habitat", 
                                             n_stations = "# stations",
                                             n_transects = "# transects",
                                             n_taxa = "# taxa",
                                             richness = "Species richness",
                                             ind_m2 = "Abundance",
                                             gr_m2 = "Biomass density")) %>% 
  flextable::merge_v(j = "sublocation") %>% 
  flextable::autofit(part = "all") %>%
  flextable::align(align = "center", part = "all")
```

```{r}
transect_summary %>% 
  group_by(depth_strata) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2, H), mean)) 
```

```{r}
transect_summary %>% 
  select(ind_m2, gr_m2, H, evenness, n_taxa) %>% 
  rstatix::shapiro_test(ind_m2, gr_m2, H, evenness, n_taxa)

adonis2(transect_summary %>% 
          select(n_taxa) ~  habitat + depth_strata + diver + sublocation, 
        data = transect_summary,
        strata = transect_summary$ps_station_id,
        by = "margin",
        method = "euclidean")
```

```{r}
var_labs <- c("A", "B", "C")

names(var_labs) <- c("richness", "gr_m2", "ind_m2")

univariate_plots_by_habitat <- transect_summary %>% 
  pivot_longer(c(richness, ind_m2, gr_m2), names_to = "variable") %>% 
  mutate(variable = factor(variable,      # Reordering group factor levels
                           levels = c("richness", "ind_m2", "gr_m2"))) %>% 
  ggplot()+
  geom_boxplot(aes(x = sublocation, y = value, fill = habitat))+
  facet_wrap("variable", scales = "free", nrow = 3, 
             labeller = labeller(variable = var_labs))+
  bbplot::bbc_style()+
  #paletteer::scale_fill_paletteer_d("ggsci::default_jama")+
  PristineSeasR::scale_fill_pristine_seas(palette = "alternative")+
  theme(legend.position = "bottom", plot.title.position = "plot") +
  labs(fill = "", x = "", title = "")+
  guides(fill=guide_legend(nrow=2,byrow=TRUE))

ggsave(univariate_plots_by_habitat, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_univariate_boxplots_by_habitat.png"), 
       dpi = 300, width = 11, height = 10)

univariate_plots_by_habitat_SP <- univariate_plots_by_habitat +
  PristineSeasR::scale_fill_pristine_seas(palette = "alternative",
                                          labels = c("Terraza prearrecifal profunda", 
                                               "Terraza prearrecifal",
                                               "Arrecife periferico", "Cuenca lagunar", 
                                               "Pastos marinos"))
  paletteer::scale_fill_paletteer_d("ggsci::default_jama", 
                                    ) 
univariate_plots_by_habitat_SP %>% 
  ggsave(filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_univariate_boxplots_by_habitat-spanish.png"), 
         dpi = 300, width = 12, height = 10)

univariate_plots_by_depth <- transect_summary %>% 
  pivot_longer(c(richness, ind_m2, gr_m2), names_to = "variable") %>% 
  ggplot()+
  geom_boxplot(aes(x = sublocation, y = value, fill = depth_strata))+
  facet_wrap("variable", scales = "free", nrow = 3, 
             labeller = labeller(variable = var_labs))+
  bbplot::bbc_style()+
  paletteer::scale_fill_paletteer_d("ggsci::default_jama")+
  theme(legend.position = "bottom", plot.title.position = "plot") +
  labs(fill = "", x = "", title = "Univariate statistics by habitat")
  
ggsave(univariate_plots_by_depth, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_univariate_boxplots_by_depth.png"), 
       dpi = 300, width = 11, height = 10)
```

```{r }
(caribe_fish_diversity_treemap <- caribe_fish_data %>% 
  group_by(sublocation, order, family) %>% 
  summarise(weight = n_distinct(ps_taxon_code))%>% 
  ggplot(aes(area = weight, 
             label = paste(family, weight, sep = "\n"),
             fill = order,
             subgroup = order)) +
  treemapify::geom_treemap(show.legend = T)+
  treemapify::geom_treemap_text(colour = "black", place = "middle", reflow = T, min.size = 3)+
  labs(fill = "",
       title = "")+
  facet_wrap("sublocation") +
  scale_fill_manual(values = colorRampPalette(RColorBrewer::brewer.pal(12, "Set3"))(24))+
  bbplot::bbc_style()+
  guides(fill = guide_legend(nrow=5,byrow=TRUE))+
  theme(legend.position = "bottom"))

ggsave(caribe_fish_diversity_treemap, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_diversity_treemap.png"), 
       dpi = 300, width = 12, height = 10)

caribe_fish_diversity_treemap_SP <- caribe_fish_diversity_treemap +
  labs(fill = "",
       title = "Diversidad de especies en censos visuales de peces")

ggsave(caribe_fish_diversity_treemap_SP, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_diversity_treemap_spanish.png"), 
       dpi = 300, width = 12, height = 10)
```

## Acc curves

```{r}
library(iNEXT)

acc_data_by_atoll <- fish_data_by_transect %>%
  mutate(presence = as.integer(if_else(n_ind == 0, 0, 1))) %>% 
  select(sublocation, transect_id, taxon_sci_name, presence) %>% 
  pivot_wider(names_from = transect_id, 
              values_from = presence, values_fill = 0) %>% 
  group_by(sublocation) %>% 
  group_split(keep = FALSE)  %>% 
  set_names(c("Bajo Nuevo", "Serranilla")) %>% 
  map(.f = ~column_to_rownames(.x, "taxon_sci_name")) %>% 
  map(.f = ~select_if(.x, ~ !is.numeric(.) || sum(.) != 0))%>% 
  map(.f = ~filter(.x, rowSums(.x)>0))

ac_curve_by_atoll <- iNEXT(acc_data_by_atoll, q = 0,  datatype = "incidence_raw",endpoint = 150)

accum_curve_by_atoll <- ggiNEXT(ac_curve_by_atoll)+
  labs(x = "", y = "")+
  theme(legend.position = "right")+
  guides(linetype = "none")+
  scale_color_manual(values = c("#4C2E05", "#5BC0BE"))+
  scale_fill_manual(values = c("#4C2E05", "#5BC0BE"))
```

```{r}
acc_data_by_habitat_serr <- fish_data_by_transect %>%
  mutate(presence = as.integer(if_else(n_ind == 0, 0, 1))) %>% 
  filter(sublocation == "Serranilla") %>% 
  select(habitat, transect_id, taxon_sci_name, presence) %>% 
  pivot_wider(names_from = transect_id, 
              values_from = presence, values_fill = 0) %>% 
  group_by(habitat) %>% 
  group_split(keep = F)  %>% 
  set_names(c("Terraza prearrecifal", "Arrecife periferico", "Cuenca lagunar", "Pastos marinos")) %>% 
  map(.f = ~column_to_rownames(.x, "taxon_sci_name")) %>% 
  map(.f = ~select_if(.x, ~ !is.numeric(.) || sum(.) != 0))%>% 
  map(.f = ~filter(.x, rowSums(.x)>0))

ac_curve_by_hab_serra <- iNEXT(acc_data_by_habitat_serr[1:3], q = 0, 
                               datatype = "incidence_raw", endpoint =100, se = T)

accum_curve_serra <- ggiNEXT(ac_curve_by_hab_serra)+
  labs(x = "", y = "Diversidad de especies")+
  theme(legend.position = "right")+
  guides(linetype = "none")+
  PristineSeasR::scale_color_pristine_seas(palette = "alternative",
                                           reverse = T)+
  PristineSeasR::scale_fill_pristine_seas(palette = "alternative",
                                          reverse = T)+
  ggtitle('Serranilla')
```

```{r}
acc_data_by_habitat_bn <- fish_data_by_transect %>%
  mutate(presence = as.integer(if_else(n_ind == 0, 0, 1))) %>% 
  filter(sublocation == "Bajo Nuevo") %>% 
  select(habitat, transect_id, taxon_sci_name, presence) %>% 
  pivot_wider(names_from = transect_id, 
              values_from = presence, values_fill = 0) %>% 
  group_by(habitat) %>% 
  group_split(keep = F)  %>% 
  set_names(c("Terraza prearrecifal profunda", "Terraza prearrecifal", "Cuenca lagunar")) %>% 
  map(.f = ~column_to_rownames(.x, "taxon_sci_name")) %>% 
  map(.f = ~select_if(.x, ~ !is.numeric(.) || sum(.) != 0)) %>% 
  map(.f = ~filter(.x, rowSums(.x)>0))

ac_curve_by_hab_bn <- iNEXT(acc_data_by_habitat_bn, q = 0, 
                               datatype = "incidence_raw", endpoint = 100,  se = T)

library(PristineSeasR)

accum_curve_bn <- ggiNEXT(ac_curve_by_hab_bn)+
  labs(x = "Número of transectos", y = "")+
  theme(legend.position = "right")+
  guides(linetype = "none")+
  PristineSeasR::scale_color_pristine_seas(palette = "alternative")+
  PristineSeasR::scale_fill_pristine_seas(palette = "alternative") +
  ggtitle('Bajo Nuevo')
```

```{r}
library(patchwork)

acc_curves <- accum_curve_by_atoll/accum_curve_serra/accum_curve_bn & 
  theme_minimal() + 
  theme(axis.text=element_text(size=12),
        axis.title=element_text(size=14,face="bold"))

ggsave(acc_curves, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_acc_curves.png"), 
       dpi = 300, width = 8, height = 10)
```

```{r}
summary_by_taxa <- fish_data_by_station %>% 
  group_by(sublocation, family, taxon_sci_name) %>% 
  summarize(across(c(n_ind, f_n_ind, ind_m2, f_ind_m2, gr_m2, f_gr_m2), mean)) %>% 
  left_join(fish_data_by_station %>% 
              group_by(sublocation, family, taxon_sci_name) %>% 
              summarize(f_obs = 100*n_distinct(ps_station_id[n_ind > 0])/n_distinct(ps_station_id))) %>% 
  mutate_if(is.numeric, round, 3) %>% 
  mutate(f_obs = round(f_obs, 1)) %>% 
  arrange(desc(gr_m2)) 
```

Fish abundance was dominated by small reef fish at both sites. In Serranilla, (*Thalassoma bifasciatum*) and (*Halichoeres bivittatus*) accounted for 25% and 7% of all fish observations. In Bajo Nuevo (*Stegastes partitus*) and (*T. bifasciatum*) were observed in all surveys and accounted for 11% and 10% of observations, on average. Most reef fish species were observed at both atolls but several others were exclusively seen on transect on only one. For instance, nurse sharks (*G. cirratum*) and (*Chaetodon capistratus*) were seen in two thirds of transects in Bajo Nuevo but were absent in Serranilla. Similarly, seven species of hamlets (*Hypoplectrus sp*) were frequently observed in Bajo Nuevo but were never seen in Serranilla. Conversely, (*Halichoeres poeyi*) and (*Haemulon parra*) and (*Pomacanthus arcuatus*) were observed in 27%, 18%, and 18% of transects in Serranilla but were absent in Bajo Nuevo. Similarly to species richness, fish abundance was significantly lower in seagrass beds that in other habitats. Including only the common habitats between atolls, average fish abundance was higher in Serranilla (2.12 ind/m2) than Bajo Nuevo (2.02 ind/m2) but differences are not statistically significant. Fish abundance was lower is deeper transects than in shallower ones and while differences are statistically significant (p = 0.017), depth explains a modest amount of the variability (R2 = 0.065)

```{r}
exclusive_sp <- summary_by_taxa %>% 
  select(sublocation, family, taxon_sci_name, f_obs) %>% 
  pivot_wider(names_from = c(sublocation), values_from = f_obs) %>% 
  filter(Serranilla == 0 | `Bajo Nuevo` == 0) %>% 
  arrange(Serranilla)

transect_summary %>% 
  filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
  group_by(sublocation) %>% 
  summarize(across(c(richness = n_taxa, n_ind, ind_m2, gr_m2, H), mean)) 

adonis2(transect_summary %>% 
          filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          select(ind_m2) ~  habitat + depth_strata + diver + sublocation, 
        data = transect_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")) ,
        strata = transect_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          pull(ps_station_id),
        by = "margin",
        method = "euclidean")
```

```{r}
adonis2(station_summary %>% 
          filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          select(ind_m2) ~  habitat + sublocation, 
        data = station_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")),
        by = "margin",
        method = "euclidean")

pairwiseAdonis::pairwise.adonis2(station_summary %>% 
                                   filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
                                   select(ind_m2) ~  habitat + sublocation, 
                                 data = station_summary %>% 
                                   filter(!habitat %in% c("deep platform", "seagrass beds")),
                                 by = "margin",
                                 method = "euclidean")
```

In Serranilla, the highest average biomass density (gr/m2) was attributed to (*Haemulon carbonarium*), (*Melichthys niger*), and (*Balistes vetula*) with values of 11.1, 8.16, and 5.32 gr/m2, respectively. These species were also very common and were observed in 82%, 91%, and 91% of all transects. In Bajo Nuevo, biomass density was dominated by nurse sharks (*Ginglymostoma cirratum*) with an average of 37.2 gr/m2. Nurse sharks were observed in two thirds of all stations and on average accounted for 28.9% of biomass in a transect. After nurse sharks, (*M. niger*) and (*Sphyraena barracuda*) contributed most in terms of biomass with average densities of 10.7 and 3.7 gr/m2, respectively. Overall, average fish biomass was slightly higher in Bajo Nuevo (91.73 gr/m2) than in Serranilla (81.78 gr/m2). However, if we omit nurse sharks from the analysis and limit the comparison to shared habitats, fish biomass density is considerably higher in Serranilla (89.43) than in Bajo Nuevo (50.72), although not statistically significant. Importantly, the biomass density of predatory fish and commercially important species (i.e., *Lutjanus* sp., *"Mycteroperca* sp., *Caranx* sp., *Seriola* sp.,) was very low at both Serranilla (2.12 gr/m2) and Bajo Nuevo (0.62 gr/m2). 

```{r}
fish_data_by_transect %>% 
  filter(family != "Ginglymostomatidae" & taxon_sci_name != "Decapterus macarellus") %>% 
    filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
  group_by(sublocation, ps_station_id, transect_id, habitat, depth_strata, depth_m, diver) %>% 
  summarize(n_taxa = n_distinct(ps_taxon_code[n_ind > 0]),
            n_ind = sum(n_ind),
            ind_m2 = sum(ind_m2),
            gr_m2 = sum(gr_m2)) %>% 
  ungroup() %>% 
  group_by(sublocation, habitat, ps_station_id) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2), mean)) %>% 
  mutate_if(is.numeric, round, 2) %>% 
  ungroup() %>% 
  group_by(sublocation) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2), mean)) 
```

```{r}
adonis2(transect_summary %>% 
          filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          select(gr_m2) ~  habitat + depth_strata + diver + sublocation, 
        data = transect_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")) ,
        strata = transect_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          pull(ps_station_id),
        by = "margin",
        method = "euclidean")

adonis2(station_summary %>% 
          filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
          select(gr_m2) ~  habitat + sublocation, 
        data = station_summary %>% 
            filter(!habitat %in% c("deep platform", "seagrass beds")),
        by = "margin",
        method = "euclidean")

adonis2(transect_summary %>% 
          filter(!habitat %in% c("seagrass beds", "deep platform")) %>% 
          select(gr_m2, ind_m2, n_taxa) ~ sublocation + habitat + depth_m + diver, 
        data = transect_summary %>% 
          filter(!habitat %in% c("seagrass beds", "deep platform")),
        strata = transect_summary$ps_station_id[!transect_summary$habitat %in% 
                                                  c("seagrass beds", "deep platform")],
        by = "margin")
```

```{r, include = T, echo = F}
summary_by_taxa %>% 
  select(sublocation, family, taxon_sci_name, n_ind, gr_m2, f_obs) %>% 
  group_by(sublocation) %>% 
  slice_max(gr_m2, n = 10) %>% 
  ungroup() %>% 
  pivot_longer(-c(sublocation, taxon_sci_name, family)) %>% 
  pivot_wider(names_from = c(sublocation, name), values_from = value) %>% 
  arrange(family, taxon_sci_name) %>% 
  flextable::flextable() %>% 
  flextable::set_header_labels(values = list(family = "Family", 
                                             taxon_sci_name = "Taxa", 
                                             `Bajo Nuevo_n_ind` = "Abundance",
                                             `Bajo Nuevo_gr_m2` = "Biomass density",
                                             `Bajo Nuevo_f_obs` = "% Ocurrence",
                                             `Serranilla_n_ind` = "Abundance",
                                             `Serranilla_gr_m2` = "Biomass density",
                                             `Serranilla_f_obs` = "% Ocurrence")) %>% 
  flextable::add_header_row(values = c("", "", "Bajo Nuevo", "Serranilla"), 
                            colwidths = c(1,1,3,3)) %>% 
  flextable::merge_v(j = "family") %>% 
  flextable::autofit(part = "all") %>%
  flextable::align(align = "center", part = "all")
```

```{r}
caribe_fish_biomass_treemap <- fish_data_by_transect %>% 
  #filter(family != "Ginglymostomatidae") %>% 
  group_by(sublocation, ps_station_id, ps_taxon_code, order, family) %>% 
  summarise(across(c(ind_m2, gr_m2), mean)) %>%
  ungroup() %>% 
  group_by(sublocation, order, family) %>% 
  summarise(across(c(ind_m2, gr_m2), sum)) %>% 
  ungroup() %>% 
  mutate(f_ind_m2 = round(100*ind_m2/sum(ind_m2),1),
         f_gr_m2 = round(100*gr_m2/sum(gr_m2), 1)) %>% 
  ggplot(aes(area = gr_m2, 
             label = paste(family, paste0(f_gr_m2, "%"), sep = "\n"),
             fill = order,
             subgroup = order)) +
  treemapify::geom_treemap(show.legend = T)+
  treemapify::geom_treemap_text(colour = "white", place = "middle", reflow = T, min.size = 3)+
  labs(fill = "",
       title = "Taxa biomass (gr m2) in underwater fish surveys")+
  facet_wrap("sublocation", nrow = 1) +
  scale_fill_manual(values = colorRampPalette(RColorBrewer::brewer.pal(12, "Set3"))(24))+
  bbplot::bbc_style()+
  guides(fill = guide_legend(nrow=5,byrow=TRUE))+
  theme(legend.position = "bottom")

ggsave(caribe_fish_biomass_treemap, 
       filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_biomass_treemap.png"),
       dpi = 300, width = 12, height = 10)
```

```{r}
fish_data_by_transect %>% 
  filter(str_detect(taxon_sci_name, c("Caranx", "Mycteroperca", "Seriola", "Lutjanus"))) %>% 
  group_by(sublocation, ps_station_id, transect_id, habitat, depth_strata, depth_m, diver) %>% 
  summarize(n_taxa = n_distinct(ps_taxon_code[n_ind > 0]),
            n_ind = sum(n_ind),
            ind_m2 = sum(ind_m2),
            gr_m2 = sum(gr_m2)) %>% 
  ungroup() %>% 
  group_by(sublocation, habitat, ps_station_id) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2), mean)) %>% 
  mutate_if(is.numeric, round, 2) %>% 
  ungroup() %>% 
  group_by(sublocation) %>% 
  summarize(across(c(n_taxa, n_ind, ind_m2, gr_m2), mean)) 
```

## Community composition

Fish assemblages were significantly different between atolls (R2 = 0.16; F = 4.75; p= 0.001). Stations in Serranilla show a higher concordance in the ordination space relative to Bajo Nuevo which is expected given the smaller sampling footprint in the latter atoll. Species that strongly determine the ordination of stations include the nurse sharks (*G. cirratum*) and hamlet (*H. nigricans*) which were abundant in Bajo Nuevo but absent in Serranilla, and (*Balistes vetula*) whose biomass density was twice as high in Serranilla compared to Bajo Nuevo. 

```{r}
gr_m2_data_wide <- fish_data_by_station %>% 
  filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
  select(ps_station_id, taxon_sci_name, gr_m2) %>% 
  pivot_wider(names_from = taxon_sci_name, values_from = gr_m2, values_fill = 0) %>% 
  column_to_rownames("ps_station_id") 

gr_m2_dist <- vegan::vegdist(as.matrix(gr_m2_data_wide), 
                      method = "bray") 

gr_m2_pcoa_c <- dbrda(gr_m2_dist ~ habitat + sublocation, 
                      data = station_summary %>% 
                        filter(!habitat %in% c("deep platform", "seagrass beds")),
                      dist = "bray", 
                      add = "lingoes")

gr_m2_pcoa_c_env_fit <- envfit(gr_m2_pcoa_c, 
                               station_summary %>% 
                                 filter(!habitat %in% c("deep platform", "seagrass beds")) %>% 
                                 select(habitat,  sublocation))

gr_m2_pcoa_c_spp_fit <- envfit(gr_m2_pcoa_c, 
                               gr_m2_data_wide)

gr_m2_pcoa_c_spp_scores <- gr_m2_pcoa_c_spp_fit %>% 
  scores("vectors") %>% 
  as_tibble(rownames = "ps_taxon_code") %>% 
  mutate(r = gr_m2_pcoa_c_spp_fit$vectors$r,
         p = gr_m2_pcoa_c_spp_fit$vectors$pvals) %>% 
  filter(p < 0.05, r > 0.1)

gr_m2_pcoa_c_axis <- BiodiversityR::axis.long(gr_m2_pcoa_c, choices = c(1, 2))

gr_m2_pcoa_c_site_scores <- scores(gr_m2_pcoa_c)$sites %>% 
  as_tibble(rownames = "ps_station_id") %>% 
  inner_join(station_summary)
```

```{r}
community_pcoa_plots <- map(c("habitat", "sublocation"), 
    ~ ggplot() +
      geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
      geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +  
      scale_x_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
      scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
      geom_point(data = gr_m2_pcoa_c_site_scores,
                 aes(x = dbRDA1, y = dbRDA2,, col = .data[[.x]])) +
      labs(x = gr_m2_pcoa_c_axis$label[1], y = gr_m2_pcoa_c_axis$label[2])+
      geom_segment(data = gr_m2_pcoa_c_spp_scores %>% 
                     filter(p < 0.05, r > 0.5), 
                   aes(x = 0, y = 0, xend = dbRDA1*3, yend = dbRDA2*3, alpha = r), 
                   colour = "black", 
                   size = 0.2, 
                   arrow = arrow(length = unit(0.01, "npc"), 
                                 type = 'open', ends = "last"))+
      ggrepel::geom_text_repel(data = gr_m2_pcoa_c_spp_scores %>% 
                                 filter(p < 0.05, r > 0.5), 
                               aes(x=dbRDA1*3, y=dbRDA2*3, label = ps_taxon_code, alpha = r),
                               colour="black",
                               show.legend = F) +
      coord_fixed(ratio=1)+
      labs(color = "", alpha = bquote('R'^'2'))+
      ggforce::geom_mark_ellipse(data = gr_m2_pcoa_c_site_scores, 
                                 aes(x=dbRDA1, y = dbRDA2, colour = .data[[.x]], 
                                     fill=after_scale(alpha(colour, 0.01))), 
                                 expand=0, size=0.2, show.legend=FALSE)+
      theme(panel.background = element_blank(),
            panel.border = element_blank(),
            panel.grid = element_blank(),
            axis.line = element_line("gray25"),
            text = element_text(size = 12, family = "Arial"),
            axis.text = element_text(size = 10, colour = "gray25"),
            axis.title = element_text(size = 12, colour = "gray25"),
            legend.key = element_blank())+
      PristineSeasR::scale_color_pristine_seas(palette = "alternative")+
      scale_alpha(range=c(0.6,1)))

pcoa_by_hab <- community_pcoa_plots[[1]] +
  labs(color = "Habitat") + 
  PristineSeasR::scale_color_pristine_seas(palette = "alternative",
                                           labels = c("Terraza prearrecifal", 
                                                "Arrecife periferico","Cuenca lagunar"))
 
pcoa_by_location <- community_pcoa_plots[[2]]+ 
  scale_color_manual(values = c("#4C2E05", "#5BC0BE"))
  
library(patchwork)

pcoa_plots <- pcoa_by_location/pcoa_by_hab 
  
ggsave(pcoa_plots, filename = file.path(ps_exp_path, 
                                        "figures", "uvc_caribe_fish_pcoa_plots.png"), 
       dpi = 300, width = 8, height = 10)

# ggsave(pcoa_by_hab, 
#        filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_biomas_pcoa_by_habitat_spanish.png"), 
#        dpi = 300, width = 8, height = 6)
# 
# 
# pcoa_by_location <- community_pcoa_plots[[2]] 
# 
#   ggsave(filename = file.path(ps_exp_path, "figures", "uvc_caribe_fish_biomas_pcoa_by_sublocation.png"), 
#        dpi = 300, width = 8, height = 6)
```

```{r}
anova(gr_m2_pcoa_c) # overall test of the significant of the analysis
anova(gr_m2_pcoa_c, by = "axis", perm.max = 500) # test axes for significance
anova(gr_m2_pcoa_c, by = "terms", perm.max = 1000) # test for sign. environ. variables
```

```{r}
n_ind_adonis <- adonis2(gr_m2_dist ~  habitat  + sublocation,
        data = station_summary[!station_summary$habitat %in% c("deep platform", "seagrass beds"), ],
        #strata = transect_summary[transect_summary$n_ind > 10, ]$ps_station_id,
        by = "margin")

n_ind_adonis
```
