==> scripts/13_analyze_meso.Rmd <==
---
title: "Process pelagic cams"
author: 
output:
  word_document:
    toc: no
---

```{r echo = FALSE, include = FALSE, warning = FALSE,prompt = FALSE, error = FALSE, strip.white = TRUE}
knitr::opts_chunk$set(echo = FALSE, include = FALSE, warning = FALSE,
                      prompt = FALSE, error = FALSE, strip.white = TRUE)

knitr::opts_knit$set(progress = TRUE, verbose = TRUE)

options(dplyr.summarise.inform = FALSE)

library(tidyverse)
library(flextable)
library(raster)
library(sf)
library(ggspatial)
library(lubridate)
library(bdc)
library(terra)
source("_prj_vars.R")

knitr::opts_chunk$set(echo = TRUE, fig.cap = TRUE)
  set_flextable_defaults(
  font.family = "Arial", font.size = 9, 
  theme_fun = "theme_vanilla",
  big.mark="", table.layout="autofit")
  
load(file = file.path(ps_exp_path, "data", "secondary", "processed", "admin_spatial_layers.RData"))

meso_maxN_caribe <- readxl::read_xlsx(file.path(ps_exp_path, 
                                    "data/primary/raw",
                                    "meso","Mesophotic_Caribbean_MaxN.xlsx")) %>% 
  janitor::clean_names() 

meso_maxN_caribe$location[meso_maxN_caribe$location == "Baja Nuevo"] <- "Bajo Nuevo"

meso_maxN_caribe$location[meso_maxN_caribe$location == "Seranilla"] <- "Serranilla"

meso_maxN_caribe <- meso_maxN_caribe %>% 
  mutate(bruvs_number = as.numeric(str_remove_all(bruvs_id, "[A-Z]+|_|[a-z]"))) %>% 
  mutate(ps_station_id = paste0("COL_meso_", 
                               formatC(bruvs_number + 6 , width = 2, flag = 0))) %>% 
  dplyr::select(ps_station_id, everything()) %>% 
  dplyr::select(-bruvs_number, -bruvs_id)

meso_meta <- read_csv(file.path(ps_exp_path, 
                               "data", "primary", "processed", "metadata", "meso_meta.csv")) %>% 
  filter(location == "Caribbean")

meso_taxa <- read_csv(file.path(ps_exp_path, 
                               "data", "primary", "processed", "taxa", "clean_master_taxa.csv")) %>% 
  filter(method == "meso", location == "Caribbean")

meso_meta <- meso_meta %>% filter(depth_m > 20)
```

```{r}
meso_maxN_caribe_long <- meso_maxN_caribe %>% 
  filter(ps_station_id %in% meso_meta$ps_station_id) %>% 
  dplyr::select(-sharks, -jacks, -snappers, -groupers, -grunts) %>% 
  pivot_longer(!c(location, ps_station_id), names_to = "taxa", values_to = 'maxN', values_drop_na = T) %>% 
  mutate(taxa = str_to_sentence(str_replace_all(taxa, "_", " "))) %>% 
  group_by(location, ps_station_id, taxa) %>% 
  summarize(max_N = max(maxN)) %>% 
  ungroup()

meso_summary_by_station <- meso_maxN_caribe_long %>%  
  group_by(location, ps_station_id) %>% 
  summarize(n_taxa = n_distinct(taxa[max_N > 0 ]),
            abundance = sum(max_N)) %>% 
  left_join(meso_meta %>% 
              group_by(ps_station_id) %>% 
              summarize(lat = lat_in,
                        lon = lon_in,
                        depth_m = depth_m)) %>% 
  ungroup()

meso_summary_by_taxa <- meso_maxN_caribe_long %>% 
  group_by(location) %>% 
  mutate(n_stations = n_distinct(ps_station_id)) %>% 
  ungroup() %>% 
  group_by(location, taxa) %>% 
  summarize(abundance = sum(max_N),
            avg_maxN = round(mean(max_N),1),
            max_maxN  = max(max_N),
            f_obs = round(100*n_distinct(ps_station_id[max_N>0])/mean(n_stations))) %>% 
  arrange(desc(f_obs)) %>% 
  ungroup()
```

We deployed a total of `r n_distinct(meso_meta$ps_station_id[meso_meta$sublocation == "Bajo Nuevo"])` mesophotic camera systems in Bajo Nuevo and `r n_distinct(meso_meta$ps_station_id[meso_meta$sublocation == "Serranilla"])` in Serranilla. Each baited camera recorded for two hours at depths between `r min(meso_meta$depth_m, na.rm = T)` and `r max(meso_meta$depth_m, na.rm = T)` meters. A total of `r n_distinct(meso_taxa$taxon)` distinct taxa belonging to `r n_distinct(meso_taxa$family)` families were recorded and some of the most frequently observed taxa include Queen triggerfish (*Balistes vetula*), White margate (*Haemulon album*), Bar jack (*C. ruber*), Caribbean reef shark (*C. perezii*), Yellowfin grouper (*Mycteroperca venenosa*), Blackfin snapper (*Lutjanus buccanella*), and Yellowtail snapper (*Ocyurus chrysurus*). We don’t find statistically significant differences in total abundance or species richness between Serranilla and Bajo Nuevo. Nonetheless the median abundance of snappers in Serranilla was six times higher than in Bajo Nuevo. 

```{r, include = T, echo = F}
meso_summary_by_taxa %>% 
  pivot_longer(-c(location, taxa)) %>% 
  pivot_wider(names_from = c(location, name), values_from = value) %>% 
  left_join(meso_taxa %>% 
              dplyr::select(taxa = taxon, family)) %>% 
  dplyr::select(family, everything()) %>% 
  arrange(family, taxa) %>% 
  flextable()  %>% 
  set_header_labels(values = list(family = "Family", 
                                  taxa = "Taxa", 
                                  Serranilla_abundance = "Abundance",
                                  `Bajo Nuevo_abundance` = "Abundance",
                                  `Bajo Nuevo_avg_maxN` = "Mean max N",
                                  `Serranilla_avg_maxN` = "Mean max N",
                                  `Bajo Nuevo_max_maxN` = "Max N",
                                  `Serranilla_max_maxN` = "Max N",
                                  `Bajo Nuevo_f_obs` = "% Ocurrence",
                                  `Serranilla_f_obs` = "% Ocurrence")) %>% 
  flextable::add_header_row(values = c("", "", "Serranilla", "Bajo Nuevo"), 
                            colwidths = c(1,1,4,4)) %>% 
  merge_v(j = "family") %>% 
  autofit(part = "all") %>%
  flextable::align(align = "center", part = "all")
```

# Maps 

```{r, eval = F}
caribe_sat_img <- terra::rast(file.path(ps_exp_path, 
                               "data", "secondary", "raw", 
                               "Allen atlas", "Coral-Reefs-2020-Visual-V1-Mosaic", "satellite_imagery_0.tif")) 

bajos_eez <- bajos_eez %>% 
  st_transform(crs(caribe_sat_img))
  
bn_sat_img <- caribe_sat_img %>% 
  terra::crop(bajos_eez %>% 
                filter(territory1 == "Bajo Nuevo Bank") %>% 
                terra::ext(), 
              mask = T) 

serr_sat_img <- caribe_sat_img %>% 
  terra::crop(bajos_eez %>% 
                filter(territory1 == "Serranilla Bank") %>% 
                ext(), 
              mask = T) 

serr_map <- meso_summary_by_station %>% 
  filter(location == "Serranilla") %>% 
  filter(!is.na(lon)) %>% 
  sf::st_as_sf(coords = c("lon", "lat"), crs = 4326) %>% 
  ggplot()+
  tidyterra::geom_spatraster_rgb(data = serr_sat_img, maxcell = 10^7)+
  geom_sf(aes(size = abundance, col = n_taxa))+
  coord_sf(xlim = c(-79.9, -79.7), ylim = c(15.72, 15.9))+
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.2, 
                              style = "ticks", 
                              line_col = "white",
                              text_col = "white",
                              pad_y = unit(0.01, units =  "native"), 
                              pad_x = unit(0.01, units =  "native"))+
  ggspatial::annotation_north_arrow(location = "tr", 
                                    which_north = "true", 
                                    height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.01, "native"), pad_y = unit(0.01, "native"),
                                    style = north_arrow_fancy_orienteering)+
  viridis::scale_color_viridis()+
  scale_shape_manual(values= seq(100:110), name = "")+
  theme_light()+
  labs(col = "# Taxa", shape = "", size = "Total max N")+
  theme(legend.position = "right",
        legend.background = element_blank(),
        legend.key = element_blank(),
        legend.text = element_text(colour = "black"), 
        panel.background  = element_rect(fill ="#1b2b49"),
        panel.grid = element_blank())

ggsave(serr_map,
       filename = file.path(ps_exp_path, "figures", "meso_serranilla_map.png"), 
              height = 5, width = 6,
       dpi = 300)

bn_map <- meso_summary_by_station %>% 
  filter(location == "Bajo Nuevo") %>% 
  filter(!is.na(lon)) %>% 
  sf::st_as_sf(coords = c("lon", "lat"), crs = 4326) %>% 
  ggplot()+
  tidyterra::geom_spatraster_rgb(data = bn_sat_img, maxcell = 10^7)+
  geom_sf(aes(size = abundance, col = n_taxa))+
  coord_sf(xlim = c(-78.8, -78.5), ylim = c(15.75, 15.95))+
  ggspatial::annotation_scale(location = "bl", 
                              width_hint = 0.2, 
                              style = "ticks", 
                              line_col = "white",
                              text_col = "white",
                              pad_y = unit(0.01, units =  "native"), 
                              pad_x = unit(0.01, units =  "native"))+
  ggspatial::annotation_north_arrow(location = "tr", 
                                    which_north = "true", 
                                    height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.01, "native"), pad_y = unit(0.01, "native"),
                                    style = north_arrow_fancy_orienteering)+
  viridis::scale_color_viridis()+
  scale_shape_manual(values= seq(100:110), name = "")+
  theme_light()+
  labs(col = "# Taxa", shape = "", size = "Total max N")+
  theme(legend.position = "right",
        legend.background = element_blank(),
        legend.key = element_blank(),
        legend.text = element_text(colour = "black"), 
        panel.background  = element_rect(fill ="#1b2b49"),
        panel.grid = element_blank())

ggsave(bn_map,
       filename = file.path(ps_exp_path, "figures", "meso_bajo_nuevo_map.png"), 
       height = 4, width = 6,
       dpi = 300)
```

# Stats

```{r}
meso_maxN_caribe %>% 
  filter(ps_station_id %in% meso_meta$ps_station_id) %>% 
  dplyr::select(location, sharks, jacks, snappers, groupers, grunts) %>% 
  pivot_longer(!location) %>% 
  ggplot()+
  geom_boxplot(aes(x = location, y = value, col = name))

meso_maxN_caribe %>% 
  filter(ps_station_id %in% meso_meta$ps_station_id) %>% 
  dplyr::select(location, sharks, jacks, snappers, groupers, grunts) %>% 
  pivot_longer(!location) %>% 
  group_by(location, name) %>% 
  summarise_all(list("median" = median, "mean" = mean, "sd" = sd)) %>% 
  arrange(name)
```

## univariate

```{r}
meso_summary_by_station %>% 
  dplyr::select(location, n_taxa, abundance) %>% 
  pivot_longer(!location) %>% 
  ggplot()+
  geom_boxplot(aes(location, value))+
  facet_wrap("name", scales = "free")+
  labs(x = "", y = "")
```

```{r}
vegan::adonis2(meso_summary_by_station %>% 
          dplyr::select(n_taxa) ~  location + depth_m, 
        data = meso_summary_by_station,
        by = "margin",
        method = "euclidean")

vegan::adonis2(meso_summary_by_station %>% 
          dplyr::select(abundance) ~  location + depth_m, 
        data = meso_summary_by_station,
        by = "margin",
        method = "euclidean")
```

## Community

```{r}
meso_maxN_wide <- meso_maxN_caribe_long %>% 
  pivot_wider(names_from = "taxa", values_from = "max_N", values_fill = 0) %>% 
  left_join(meso_summary_by_station %>% 
              dplyr::select(ps_station_id, depth_m))

max_n_dist <- meso_maxN_wide %>% 
  dplyr::select(-location, -depth_m) %>% 
  column_to_rownames("ps_station_id") %>% 
  vegan::avgdist(dmethod = "bray", sample = 1) 
```

```{r}
meso_pcoa_c <- vegan::dbrda(max_n_dist ~ location + depth_m, 
                            data = meso_maxN_wide,
                            dist = "bray", 
                            add = "lingoes")

meso_pcoa_c_env_fit <- vegan::envfit(meso_pcoa_c, 
                                     meso_maxN_wide %>% 
                                       dplyr::select(location, depth_m))

meso_pcoa_c_spp_fit <- vegan::envfit(meso_pcoa_c, 
                                     meso_maxN_wide %>% 
                                       dplyr::select(-location, -depth_m, -ps_station_id))

meso_pcoa_c_spp_scores <- meso_pcoa_c_spp_fit %>% 
  vegan::scores("vectors") %>% 
  as_tibble(rownames = "ps_taxon_code") %>% 
  mutate(r = meso_pcoa_c_spp_fit$vectors$r,
         p = meso_pcoa_c_spp_fit$vectors$pvals) 

meso_pcoa_c_axis <- BiodiversityR::axis.long(meso_pcoa_c, choices = c(1, 2))

meso_pcoa_c_site_scores <- vegan::scores(meso_pcoa_c)$sites %>% 
  as_tibble(rownames = "ps_station_id") %>% 
  left_join(meso_summary_by_station)
```

```{r}
anova(meso_pcoa_c) # overall test of the significant of the analysis
anova(meso_pcoa_c, by = "axis", perm.max = 500) # test axes for significance
anova(meso_pcoa_c, by = "terms", perm.max = 10000) # test for sign. environ. variables
```

```{r}
meso_pcoa_plot <- ggplot() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +  
  scale_x_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  geom_point(data = meso_pcoa_c_site_scores,
             aes(x = dbRDA1, y = dbRDA2, col = location)) +
  labs(x = meso_pcoa_c_axis$label[1], y = meso_pcoa_c_axis$label[2])+
  geom_segment(data = meso_pcoa_c_spp_scores %>% 
                             filter(r > 0.33), 
               aes(x = 0, y = 0, xend = dbRDA1*3, yend = dbRDA2*3, alpha = r), 
               colour = "black", 
               size = 0.2, 
               arrow = arrow(length = unit(0.01, "npc"), 
                             type = 'open', ends = "last"))+
  ggrepel::geom_text_repel(data = meso_pcoa_c_spp_scores %>% 
                             filter(r > 0.33), 
                           aes(x=dbRDA1*3, y=dbRDA2*3, label = ps_taxon_code, alpha = r),
                           colour="black",
                           show.legend = F) +
  coord_fixed(ratio=1)+
  ggforce::geom_mark_ellipse(data = meso_pcoa_c_site_scores, 
                             aes(x=dbRDA1, y = dbRDA2, colour = location, 
                                 fill=after_scale(alpha(colour, 0.01))), 
                             expand=0, size=0.2, show.legend=FALSE)+
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        panel.grid = element_blank(),
        axis.line = element_line("gray25"),
        text = element_text(size = 12, family = "Arial"),
        axis.text = element_text(size = 10, colour = "gray25"),
        axis.title = element_text(size = 12, colour = "gray25"),
        legend.key = element_blank())+
  scale_alpha(range=c(0.6,1))+
  scale_color_manual(values = c("#4C2E05", "#5BC0BE"))+
  #paletteer::scale_color_paletteer_d("ggsci::default_jama")+
  labs(alpha = bquote('R'^'2'), color = "")
  
ggsave(meso_pcoa_plot,
       filename = file.path(ps_exp_path, "figures", "meso_caribe_pcoa_plot.png"),   
       dpi = 300, width = 6, height = 4)
```

```{r}
meso_MDS <- vegan::metaMDS(max_n_dist, k = 2, trymax = 100)

meso_MDS$stress

meso_MDS_spp_fit <- vegan::envfit(meso_MDS, 
                                  meso_maxN_wide %>% 
                                    dplyr::select(-location, -ps_station_id, -depth_m))

meso_MDS_spp_scores <- meso_MDS_spp_fit %>% 
  vegan::scores("vectors") %>% 
  as_tibble(rownames = "ps_taxon_code") %>% 
  mutate(r = meso_MDS_spp_fit$vectors$r,
         p = meso_MDS_spp_fit$vectors$pvals) 

n_ind_MDS_site_scores <- vegan::scores(meso_MDS, display = "sites") %>% 
  as_tibble(rownames = "ps_station_id") %>% 
  left_join(meso_maxN_wide %>%
              dplyr::select(location, ps_station_id, depth_m))
```

```{r}
ggplot() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +  
  scale_x_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  scale_y_continuous(sec.axis = dup_axis(labels=NULL, name=NULL)) +
  geom_point(data = n_ind_MDS_site_scores,
             aes(x = NMDS1, y = NMDS2, col = location)) +
  geom_segment(data = meso_MDS_spp_scores %>% 
                 filter(r > 0.3), 
               aes(x = 0, y = 0, xend = NMDS1*1, yend = NMDS2*1, alpha = r), 
               colour = "black", 
               size = 0.2, 
               arrow = arrow(length = unit(0.01, "npc"), 
                             type = 'open', ends = "last"))+
  ggrepel::geom_text_repel(data = meso_MDS_spp_scores %>% 
                 filter(r > 0.3), 
                           aes(x=NMDS1*1, y=NMDS2*1, label = ps_taxon_code, alpha = r),
                           colour="black",
                           show.legend = F) +
  coord_fixed(ratio=1)+
  ggforce::geom_mark_ellipse(data = n_ind_MDS_site_scores, 
                             aes(x=NMDS1, y = NMDS2, colour = location, 
                                 fill=after_scale(alpha(colour, 0.01))), 
                             expand=0, size=0.2, show.legend=FALSE)+
  theme(panel.background = element_blank(),
        panel.border = element_blank(),
        panel.grid = element_blank(),
        axis.line = element_line("gray25"),
        text = element_text(size = 12, family = "Arial"),
        axis.text = element_text(size = 10, colour = "gray25"),
        axis.title = element_text(size = 12, colour = "gray25"),
        legend.key = element_blank())+
  scale_alpha(range=c(0.6,1))
```

```{r}
meso_adonis <- vegan::adonis2(max_n_dist ~ location + depth_m,
        data = meso_maxN_wide,
        by = "margin")

meso_adonis
```
Our data suggest there are differences in mesophotic community composition species between both atolls (R2 = 0.15074; F=1.61, p-value=0.009). The taxa that correlate strongest with the ordination of sites are (*C. lugubris*) and (*Sphyraena barracuda*) both of which were considerably more common in Bajo Neuvo than in Serranilla. Conversely, as mentioned above, snappers (i.e. *Lutjanus buccanella*) were more abundant in Serranilla. 

```{r}
meso_pcoa_c_spp_scores %>% 
  arrange(desc(r))
```

# SI

```{r include = T, echo = F}
meso_summary_by_station %>% 
  mutate_if(is.numeric, round, 3) %>% 
  left_join(meso_meta %>% 
              dplyr::select(date, ps_station_id, time = time_in)) %>% 
  dplyr::select(date, location, ps_station_id, lat, lon, time, depth_m) %>% 
  flextable() %>% 
  merge_v(j = "family") %>% 
  autofit(part = "all") %>%
  flextable::align(align = "center", part = "all")
```
